requirements.txt
click==7.1.1
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
Werkzeug==1.0.0
psycopg2-binary==2.8.6
PyJWT==1.4.2
flask-jwt==0.3.2
wordcloud==1.3.1
tweepy == 3.10.0
